add TaskResponseDTO e TaskRequestDTO

+++ src/main/java/com/junyor/taskmanager/service/TaskService.java
@@ -1,10 +1,15 @@
 package com.junyor.taskmanager.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.junyor.taskmanager.dto.TaskRequestDTO;
+import com.junyor.taskmanager.dto.TaskResponseDTO;
 import com.junyor.taskmanager.model.Task;
 import com.junyor.taskmanager.repository.TaskRepository;
 
@@ -13,39 +18,87 @@ public class TaskService {
 
     private final TaskRepository taskRepository;
 
-    public TaskService(TaskRepository taskRepository){
+    public TaskService(TaskRepository taskRepository) {
         this.taskRepository = taskRepository;
 
     }
-    
-    public Optional<Task> findById(Long id){
+
+    public Optional<Task> findById(Long id) {
         return taskRepository.findById(id);
     }
-    
-    public List<Task> getAllTasks() {
-        return taskRepository.findAll();
+
+    public List<TaskResponseDTO> getAllTasks() {
+        List<Task> tasks = taskRepository.findAll();
+        return tasks.stream()
+                .map(task -> new TaskResponseDTO(
+                        task.getId(),
+                        task.getTitle(),
+                        task.isCompleted(),
+                        task.getDescription(),
+                        task.getCreatedAt()))
+                .toList();
     }
 
-    public Task createTask(Task task) {
-        return taskRepository.save(task);
+    public TaskResponseDTO createTask(TaskRequestDTO dto) {
+        Task task = new Task();
+        task.setTitle(dto.getTitle());
+        task.setDescription(dto.getDescription());
+        task.setCreatedAt(LocalDateTime.now());
 
+        Task saved = taskRepository.save(task);
+
+        return new TaskResponseDTO(saved.getId(), saved.getTitle(), saved.isCompleted(), saved.getDescription(),
+                saved.getCreatedAt());
     }
 
-    public List<Task> listByCompleted() {
-        return taskRepository.findByCompleted(true);
+    public List<TaskResponseDTO> listByCompleted() {
+        List<Task> tasks = taskRepository.findByCompleted(true);
+        return tasks.stream()
+                .map(task -> new TaskResponseDTO(
+                        task.getId(),
+                        task.getTitle(),
+                        task.isCompleted(),
+                        task.getDescription(),
+                        task.getCreatedAt()))
+                .toList();
     }
 
-    public List<Task> listByPending() {
-        return taskRepository.findByCompleted(false);
+    public List<TaskResponseDTO> listByPending() {
+        List<Task> tasks = taskRepository.findByCompleted(false);
+        return tasks.stream()
+                .map(task -> new TaskResponseDTO(
+                        task.getId(),
+                        task.getTitle(),
+                        task.isCompleted(),
+                        task.getDescription(),
+                        task.getCreatedAt()))
+                .toList();
     }
 
-    public void deleteTask(Long id) {
-        taskRepository.deleteById(id);
+    public Optional<Task> updateTask(Long id, Map<String, Object> updates) {
+        return taskRepository.findById(id)
+                .map(task -> {
+                    if (updates.containsKey("title")) {
+                        task.setTitle((String) updates.get("title")); // atualiza titulo se estiver no json
+                    }
+                    if (updates.containsKey("completed")) {
+                        task.setCompleted((Boolean) updates.get("completed")); // atualiza status se estiver no json
+                    }
+                    if (updates.containsKey("description")) {
+                        task.setDescription((String) updates.get("description")); // atualiza descrição se estiver no
+                                                                                  // json
+                    }
+                    return taskRepository.save(task);
+                });
+            
     }
 
-    public void updateTask(Task task) {
-        taskRepository.save(task);
+    public boolean deleteTask(Long id) {
+        if (taskRepository.existsById(id)){
+            taskRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
-
 
 }
